requirements.txt
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
fvcore>=0.1.5
scikit-image>=0.19.0
colormath>=3.0.0
opencv-python>=4.5.0
tqdm>=4.60.0
lmdb>=1.2.0
tensorboardX>=2.2
wandb>=0.12.0
clip @ git+https://github.com/openai/CLIP.git
numpy>=1.21.0
Pillow>=8.0.0
matplotlib>=3.5.0
